Updates to add date metadata to the head
Using the OpenGraph article:published_time property, we can indicate to search engines the last time the content was updated to use it for relevancy ranking.

Note that the format should be in ISO format, but I haven't figured out how accomplish this.

All pages ideally would have a last_modified date associated. One implementation is via front matter, but if there is a programmatic way to automatically update these values anytime the page content is updated that may also accomplish the same goal!